requirements.txt
Django==2.1.1
django-pandas==0.5.1
djangorestframework==3.8.2
numpy==1.15.1
pandas==0.23.4
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
